Added in timer test, modified timer
Right now the timer test in test_proc test all the required
things but it needs to be modified to use microseconds for the
timer. I have figured out the value for the PR register needs to
be 11.5 in order to get 1 microsecond timer interrupts, but they seem
to be just keeping the system at halt because of so many timer
interrupts. Will continue...