Resolves #66, #63, #62, #61, #60
Added sections for the wall clock, the set priority process, the crt, the
kcd, the keyboard interrupts, and miscellaneous explanations of each.
Needs to be converted to latex and merged into the main document. Left in
ODT format for simplification of conversion (there's an extension to
openoffice that will do this).